adding test division

+++ src/main/java/travis_ci_tutorial_java/SimpleCalculator.java
@@ -11,8 +11,8 @@ public class SimpleCalculator {
 		return a * b;
 	}
 	/*
-	public int divide(int a, int b) {
-		return a / b;
+	public double divide(int a, int b) {
+		return (double)a / (double)b;
 	}
 	*/
 }
